+++ src/main/java/forestry/apiculture/EndFlowerType.java
@@ -1,11 +1,10 @@
 package forestry.apiculture;
 
 import net.minecraft.core.BlockPos;
+import net.minecraft.tags.BiomeTags;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
-
-import forestry.api.ForestryTags;
 
 public class EndFlowerType extends FlowerType {
 	public EndFlowerType(TagKey<Block> acceptableFlowers, boolean dominant) {
@@ -14,6 +13,6 @@ public class EndFlowerType extends FlowerType {
 
 	@Override
 	public boolean isAcceptableFlower(Level level, BlockPos pos) {
-		return level.getBiome(pos).is(ForestryTags.Biomes.THE_END_CATEGORY) || super.isAcceptableFlower(level, pos);
+		return level.getBiome(pos).is(BiomeTags.IS_END) || super.isAcceptableFlower(level, pos);
 	}
 }